double->float, add tickRate in blocks, reenable gen stalactites/stalagmites, some more compact or slightly optimal changes and etc

+++ src/main/java/biomesoplenty/common/biome/BOPSubBiome.java
@@ -3,9 +3,9 @@ package biomesoplenty.common.biome;
 public class BOPSubBiome extends BOPOverworldBiome
 {
 	 /**Smaller numbers zoom in the noise field (biomes are less common)*/
-	public double zoom;
+	public float zoom;
 	/**The strength the field must reach to replace the biome. Larger numbers result in smaller patches.*/
-	public double threshold;
+	public float threshold;
 	
 	public BOPSubBiome(int biomeID) 
 	{
